v2v: Only emit fstrim warning when debugging (RHBZ#1168144).
fstrim is an optimization.  If it fails, it's not fatal (although
virt-v2v will run a lot more slowly).

We also expect that it will fail for non-aligned NTFS partitions found
on old versions of Windows, and in that case there's nothing that can
be done about it.

Therefore only emit the warning when we are debugging.

(cherry picked from commit 0576fdd0b6bafaeaa3d69bf6fc6fef0d0a1254d1)