v2v: efi: Support output of UEFI guests, for some output drivers (RHBZ#1184690).
Add the Types.target_firmware type, which stores our final decision
for whether the guest requires BIOS or UEFI on the target.

Not all output modes support UEFI.  In order to fail as early as
possible if conversion isn't going to be possible, there is an
output#supported_firmware method which returns the list of supported
target_firmware.

Add the target_firmware parameter to output#create_metadata so it can
select the correct firmware in the final metadata.

(cherry picked from commit 40558450dd87190c9dc2649cadbf284ae0a21606)